+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     * Adding more functions to the Kernel module
     * Pipe operator
     * Can now call properties and zero parameter functions correctly
+    * case, cond, and if are now turned into if statements wrapped in function closures
 
 # v0.1.0